Changed to using doubles for axis internally

+++ Terminator/Device/State.cs
@@ -4,6 +4,6 @@ namespace Terminator.Device
 {
     public class State
     {
-        public IDictionary<Axis, int> Axis { get; set; }
+        public IDictionary<Axis, double> Axis { get; set; }
     }
 }
